requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
 pydantic-settings==2.3.1
python-dotenv==1.0.1
tqdm==4.66.4
sentence-transformers==3.0.1
chromadb==0.5.3
pdfplumber==0.11.0
PyMuPDF==1.24.5
langchain-text-splitters==0.2.2
numpy==1.26.4
scikit-learn==1.5.1
docling==1.13.0 ; platform_system != "Darwin" or python_version >= "3.10"
google-generativeai==0.7.2
tenacity==8.2.3
starlette==0.37.2
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
streamlit==1.36.0

# Pinecone cloud vector DB
pinecone-client==5.0.1

# LlamaParse ingestion stack
llama-parse==0.4.4
llama-index-core==0.10.54
llama-index-readers-llama-parse==0.1.4
pypdf==5.0.0

